clean up imports initial pass of the 51 critical code smells from sonarqube wire in codeql with the sonarcloud build, see if we can run both in one task

+++ src/main/java/org/maps/network/protocol/impl/mqtt5/packet/Publish5.java
@@ -66,7 +66,7 @@ public class Publish5 extends MQTTPacket5 {
     }
 
     destinationName = readUTF8(packet);
-    int nonData = destinationName.length() + 2;
+    long nonData = destinationName.length() + 2;
 
     if (qos.isSendPacketId()) {
       nonData += 2;
